Extract UsersyncInfoAssembler as self-sufficient object
