+++ src/main/java/org/prebid/server/bidder/HttpAdapterConnector.java
@@ -29,7 +29,6 @@ import org.prebid.server.proto.response.Bid;
 import org.prebid.server.proto.response.BidderDebug;
 import org.prebid.server.proto.response.BidderStatus;
 import org.prebid.server.proto.response.MediaType;
-import org.prebid.server.proto.response.UsersyncInfo;
 import org.prebid.server.vertx.http.HttpClient;
 import org.prebid.server.vertx.http.model.HttpClientResponse;
 
@@ -216,7 +215,7 @@ public class HttpAdapterConnector {
             final Privacy privacy = privacyExtractor.validPrivacyFrom(preBidRequest.getRegs(), preBidRequest.getUser());
             bidderStatusBuilder
                     .noCookie(true)
-                    .usersync(UsersyncInfo.from(usersyncer).withPrivacy(privacy).assemble());
+                    .usersync(UsersyncInfoAssembler.from(usersyncer).withPrivacy(privacy).assemble());
         }
 
         final List<Result<List<Bid>>> bidsWithErrors = exchangeCalls.stream()
