+++ src/test/java/org/prebid/server/bidder/UsersyncInfoAssemblerTest.java
@@ -1,19 +1,19 @@
-package org.prebid.server.proto.response;
+package org.prebid.server.bidder;
 
 import org.junit.Test;
-import org.prebid.server.bidder.Usersyncer;
 import org.prebid.server.privacy.ccpa.Ccpa;
 import org.prebid.server.privacy.model.Privacy;
+import org.prebid.server.proto.response.UsersyncInfo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UsersyncInfoTest {
+public class UsersyncInfoAssemblerTest {
 
     @Test
     public void assembleUsersyncInfoShouldAppendRedirectUrlToUsersyncUrl() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url/redirect=", "redirectUrl",
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url/redirect=", "redirectUrl",
                         "http://localhost:8000", null, false)).assemble();
 
         // then
@@ -23,8 +23,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleUsersyncInfoShouldAppendEncodedRedirectUrlAndNotEncodedQueryParamsToUsersyncUrl() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url/redirect=", "/setuid?gdpr={{gdpr}}?gdpr={{gdpr}}",
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url/redirect=", "/setuid?gdpr={{gdpr}}?gdpr={{gdpr}}",
                         "http://localhost:8000", null, false)).assemble();
 
         // then
@@ -35,8 +35,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleUsersyncInfoShouldIgnoreRedirectUrlIfNotDefined() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url/redirect=", null, null, null, false)).assemble();
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url/redirect=", null, null, null, false)).assemble();
 
         // then
         assertThat(result.getUrl()).isEqualTo("http://url/redirect=");
@@ -45,8 +45,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleWithPrivacyShouldCreatePrivacyAwareUsersyncInfo() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url?redir=%26gdpr%3D{{gdpr}}"
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url?redir=%26gdpr%3D{{gdpr}}"
                         + "%26gdpr_consent%3D{{gdpr_consent}}"
                         + "%26us_privacy={{us_privacy}}",
                         null, null, null, false))
@@ -59,8 +59,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleWithPrivacyShouldTolerateMissingPrivacyParamsUsersyncInfo() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url?redir=%26gdpr%3D{{gdpr}}"
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url?redir=%26gdpr%3D{{gdpr}}"
                         + "%26gdpr_consent%3D{{gdpr_consent}}"
                         + "%26us_privacy%3D{{us_privacy}}",
                         null, null, null, false)).withPrivacy(Privacy.empty()).assemble();
@@ -72,8 +72,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleWithPrivacyShouldIgnorePrivacyParamsIfTheyAreMissingInUrl() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url?redir=a%3Db", null, null, null, false))
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url?redir=a%3Db", null, null, null, false))
                 .withPrivacy(Privacy.of("1", "consent", Ccpa.of("YNN"))).assemble();
 
         // then
@@ -83,8 +83,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleWithPrivacyUsersyncInfoShouldPopulateWithPrivacyRedirectAndUsersyncUrl() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url/{{gdpr}}/{{gdpr_consent}}?redir=",
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url/{{gdpr}}/{{gdpr_consent}}?redir=",
                         "/setuid?bidder=adnxs&gdpr={{gdpr}}&gdpr_consent={{gdpr_consent}}"
                                 + "&us_privacy={{us_privacy}}"
                                 + "&uid=$UID",
@@ -100,8 +100,8 @@ public class UsersyncInfoTest {
     @Test
     public void assembleWithUrlUsersyncInfoShouldUpdateUsersyncUrl() {
         // given and when
-        final UsersyncInfo result = UsersyncInfo.UsersyncInfoAssembler
-                .assembler(new Usersyncer(null, "http://url", null, null, null, false))
+        final UsersyncInfo result = UsersyncInfoAssembler
+                .from(new Usersyncer(null, "http://url", null, null, null, false))
                 .withUrl("http://updated-url").assemble();
 
         // then
